Added new Stan code for CMR
Each version sets the parameters in the filename to zero, to create different restricted versions of the model.  As it turns out, including the init_cf parameter makes the model not identifiable if there are no meaningful differences in pre-experimental strengths (since the resulting effect is just to scale up/down the dot products, which then gets counteracted by the gamma parameter at the decision stage).